pyproject.toml
[project]
name = "multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.51",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.12",
    "langchain-together>=0.3.0",
    "langgraph>=0.3.25",
    "langgraph-cli>=0.1.89",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
]
